change operator LK-light to use the mbp_client

+++ resources/operators/extraction/LK-light_raspberry-pi/install.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
-sudo apt-get update;
-sudo apt-get install -y python3;
-sudo apt-get install -y python3-pip;
-pip3 install paho-mqtt;
+paho_exist=$(pip3 list | grep -F paho-mqtt)
+echo $paho_exist >> install.log;
+
+if [ -z $paho_exist ]; then
+ sudo apt-get update;
+ sudo apt-get install -y python3;
+ sudo apt-get install -y python3-pip;
+ pip3 install paho-mqtt;
+fi
+# install library to read analog sensors via spi
 pip3 install spidev;
-pip3 install RPi.GPIO;
-echo "$1 = $2" > $3/connections.txt;
